onjectives system 1

+++ VTOLVR-Multiplayer/Patches/Patch_VTEventInfo.cs
@@ -77,8 +77,26 @@ class PatchGPS
         }
         return true;
     }
+} 
+
+
+[HarmonyPatch(typeof(ShipSurviveObjective), "OnDeath")]
+class deathmenu
+{
+    static bool Prefix()
+    {
+        return false;
+    }
 }
 
+[HarmonyPatch(typeof(ShipSurviveObjective), "OnDeathDelayed")]
+class deathmenu2
+{
+    static bool Prefix()
+    {
+        return false;
+    }
+}
 
 [HarmonyPatch(typeof(VTEventTarget), "Invoke")]
 class Patch22
@@ -99,6 +117,9 @@ class Patch22
                 bool shouldComplete = ObjectiveNetworker_Reciever.completeNextEvent;
                 Debug.Log($"Should complete is {shouldComplete}.");
                 ObjectiveNetworker_Reciever.completeNextEvent = false;
+                if (__instance.targetType == VTEventTarget.TargetTypes.Objective)
+                    return true;
+
                 return shouldComplete;// clients should not send kill obj packets or have them complete
 
             }
@@ -117,17 +138,24 @@ class Patch2
         {
             actionIdentifier += aparam.name;
         }*/
-        int hash = actionIdentifier.GetHashCode();
+     
         if (!__instance.TargetExists())
         {
             Debug.Log("Target doesn't exist in invoke");
         }
-        Message_ScenarioAction ScanarioActionOutMessage = new Message_ScenarioAction(PlayerManager.localUID, hash);
+       
         if (Networker.isHost)
         {
             Debug.Log("Host sent Event action" + __instance.eventName + " of type " + __instance.methodName + " for target " + __instance.targetID);
+            if (ObjectiveNetworker_Reciever.reverseScenarioActionsList.ContainsKey(__instance))
+            {
+                int hash = 0;
 
-            NetworkSenderThread.Instance.SendPacketAsHostToAllClients(ScanarioActionOutMessage, Steamworks.EP2PSend.k_EP2PSendReliable);
+                hash = ObjectiveNetworker_Reciever.reverseScenarioActionsList[__instance];
+                Message_ScenarioAction ScanarioActionOutMessage = new Message_ScenarioAction(PlayerManager.localUID, hash);
+                NetworkSenderThread.Instance.SendPacketAsHostToAllClients(ScanarioActionOutMessage, Steamworks.EP2PSend.k_EP2PSendReliable);
+            }
+                
         }
         else
         {
@@ -153,7 +181,7 @@ class Patch3
             vTEventTarget.LoadFromNode(node);
             __instance.actions.Add(vTEventTarget);
             Debug.Log("Compiling scenario dictonary my codd2");
-            String actionIdentifier = vTEventTarget.eventName + vTEventTarget.methodName + vTEventTarget.targetID + vTEventTarget.targetType.ToString();
+            String actionIdentifier = __instance.eventName + vTEventTarget.eventName + vTEventTarget.methodName + vTEventTarget.targetID + vTEventTarget.targetType.ToString();
             /*foreach(VTEventTarget.ActionParamInfo aparam in vTEventTarget.parameterInfos)
             {
                 actionIdentifier+= aparam.name;
@@ -164,8 +192,10 @@ class Patch3
 
             if (!ObjectiveNetworker_Reciever.scenarioActionsList.ContainsKey(hash))
                 ObjectiveNetworker_Reciever.scenarioActionsList.Add(hash, vTEventTarget);
-            else
-                Debug.Log("Duplicate VT scenario actions found, we should probably rewrite the dictionary code");
+
+            if(!ObjectiveNetworker_Reciever.reverseScenarioActionsList.ContainsKey(vTEventTarget))
+                ObjectiveNetworker_Reciever.reverseScenarioActionsList.Add(vTEventTarget, hash);
+            
 
         }
         return false;//dont run bahas code
@@ -183,25 +213,17 @@ class Patch4
 {
     static bool Prefix(MissionObjective __instance)
     {
-        //prevents infinite client host pings
-        //if (__instance.completed)
-        //   return false;
+    
         Debug.Log("A mission got completed we need to send it");
 
-
-        // Debug.Log("sending __instance.objectiveName + __instance.objectiveID");
-        String actionIdentifier = __instance.objectiveName + __instance.objectiveID;
-
-        Debug.Log(actionIdentifier);
-
-        //dont run corrupt objectives
-        if (MissionManager.instance.IndexOfObjective(__instance) == -1)
-            return false;
-        Message_ObjectiveSync objOutMessage = new Message_ObjectiveSync(PlayerManager.localUID, MissionManager.instance.IndexOfObjective(__instance), ObjSyncType.EMissionCompleted);
-        if (Networker.isHost && objOutMessage.objID != -1)
+        int hashCode = ObjectiveNetworker_Reciever.getObjectiveHash(__instance);
+     
+        Message_ObjectiveSync objOutMessage = new Message_ObjectiveSync(PlayerManager.localUID,hashCode, ObjSyncType.EMissionCompleted);
+        if (Networker.isHost)
         {
             Debug.Log("Host sent objective complete " + __instance.objectiveID);
             ObjectiveNetworker_Reciever.completeNext = false;
+            ObjectiveNetworker_Reciever.ObjectiveHistory.Add(objOutMessage);
             NetworkSenderThread.Instance.SendPacketAsHostToAllClients(objOutMessage, Steamworks.EP2PSend.k_EP2PSendReliable);
         }
         else
@@ -225,25 +247,16 @@ class Patch4
 class Patch5
 {
     static bool Prefix(MissionObjective __instance)
-    {
-        //prevents infinite client host pings
-        //if (__instance.failed)
-        //    return true;
+    { 
         Debug.Log("A mission got failed we need to send it");
 
+  int hashCode = ObjectiveNetworker_Reciever.getObjectiveHash(__instance);
 
-        //Debug.Log("sending __instance.objectiveName + __instance.objectiveID");
-        String actionIdentifier = __instance.objectiveName + __instance.objectiveID;
-
-        Debug.Log(actionIdentifier);
-
-        //dont run corrupt objectives
-        if (MissionManager.instance.IndexOfObjective(__instance) == -1)
-            return false;
-        Message_ObjectiveSync objOutMessage = new Message_ObjectiveSync(PlayerManager.localUID, MissionManager.instance.IndexOfObjective(__instance), ObjSyncType.EMissionFailed);
-        if (Networker.isHost && objOutMessage.objID != -1)
+    Message_ObjectiveSync objOutMessage = new Message_ObjectiveSync(PlayerManager.localUID, hashCode, ObjSyncType.EMissionFailed);
+        if (Networker.isHost)
         {
             Debug.Log("Host sent objective fail " + __instance.objectiveID);
+            ObjectiveNetworker_Reciever.ObjectiveHistory.Add(objOutMessage);
             NetworkSenderThread.Instance.SendPacketAsHostToAllClients(objOutMessage, Steamworks.EP2PSend.k_EP2PSendReliable);
         }
         else
@@ -258,32 +271,21 @@ class Patch5
         return true;
     }
 }
-
-/*
+  
 //patch to grab all the events being loaded on creation this replaces original method
 [HarmonyPatch(typeof(MissionObjective), "BeginMission")]
 class Patch6
 {
     static bool Prefix(MissionObjective __instance)
     {
-        //prevents infinite client host pings
-        //if (__instance.failed)
-        //    return true;
-        Debug.Log("A mission got failed we need to send it");
+        Debug.Log("A mission got begin we need to send it");
 
-
-        //Debug.Log("sending __instance.objectiveName + __instance.objectiveID");
-        String actionIdentifier = __instance.objectiveName + __instance.objectiveID;
-
-        Debug.Log(actionIdentifier);
-
-        //dont run corrupt objectives
-        if (MissionManager.instance.IndexOfObjective(__instance) == -1)
-            return false;
-        Message_ObjectiveSync objOutMessage = new Message_ObjectiveSync(PlayerManager.localUID, MissionManager.instance.IndexOfObjective(__instance), ObjSyncType.EMissionBegin);
-        if (Networker.isHost && objOutMessage.objID != -1)
+        int hashCode = ObjectiveNetworker_Reciever.getObjectiveHash(__instance);
+        Message_ObjectiveSync objOutMessage = new Message_ObjectiveSync(PlayerManager.localUID, hashCode, ObjSyncType.EMissionBegin);
+        if (Networker.isHost)
         {
-            Debug.Log("Host sent objective fail " + __instance.objectiveID);
+            Debug.Log("Host sent objective begin " + __instance.objectiveID);
+            ObjectiveNetworker_Reciever.ObjectiveHistory.Add(objOutMessage);
             NetworkSenderThread.Instance.SendPacketAsHostToAllClients(objOutMessage, Steamworks.EP2PSend.k_EP2PSendReliable);
         }
         else
@@ -292,12 +294,12 @@ class Patch6
             bool shouldComplete = ObjectiveNetworker_Reciever.completeNextBegin;
             Debug.Log($"Should complete is {shouldComplete}.");
             ObjectiveNetworker_Reciever.completeNextBegin = false;
-            return shouldComplete;// clients should not send kill obj packets or have them complete
+            //return shouldComplete;// clients should not send kill obj packets or have them complete
             //NetworkSenderThread.Instance.SendPacketToSpecificPlayer(Networker.hostID, objOutMessage, Steamworks.EP2PSend.k_EP2PSendUnreliable);
         }
         return true;
     }
-}*/
+}  
 
 
 //patch to grab all the events being loaded on creation this replaces original method
@@ -306,26 +308,17 @@ class Patch7
 {
     static bool Prefix(MissionObjective __instance)
     {
-        //prevents infinite client host pings
-        //if (__instance.cancelled)
-        //   return true;
-
+       
         Debug.Log("A mission got CancelObjective we need to send it");
 
+        int hashCode = ObjectiveNetworker_Reciever.getObjectiveHash(__instance);
 
-        ///Debug.Log("sending __instance.objectiveName + __instance.objectiveID");
-        String actionIdentifier = __instance.objectiveName + __instance.objectiveID;
-
-        Debug.Log(actionIdentifier);
-
-        //dont run corrupt objectives
-        if (MissionManager.instance.IndexOfObjective(__instance) == -1)
-            return false;
-        Message_ObjectiveSync objOutMessage = new Message_ObjectiveSync(PlayerManager.localUID, MissionManager.instance.IndexOfObjective(__instance), ObjSyncType.EMissionCanceled);
+        Message_ObjectiveSync objOutMessage = new Message_ObjectiveSync(PlayerManager.localUID, hashCode, ObjSyncType.EMissionCanceled);
         if (Networker.isHost && objOutMessage.objID != -1)
         {
 
             Debug.Log("Host sent objective CancelObjective " + __instance.objectiveID);
+            ObjectiveNetworker_Reciever.ObjectiveHistory.Add(objOutMessage);
             NetworkSenderThread.Instance.SendPacketAsHostToAllClients(objOutMessage, Steamworks.EP2PSend.k_EP2PSendReliable);
         }
         else
